- added projects section & restructured sections
