+++ src/App.jsx
@@ -1,6 +1,7 @@
 import './App.css'
-import { BioSection } from './sections/BioSection'
-import { SkillsSection } from './sections/SkillsSection'
+import { BioSection } from "./sections/BioSection/BioSection"
+import { SkillsSection } from "./sections/SkillsSection/SkillsSection"
+import { ProjectsSection } from "./sections/ProjectsSection/ProjectsSection"
 
 function App() {
 
@@ -8,6 +9,7 @@ function App() {
     <>
       <BioSection />
       <SkillsSection />
+      <ProjectsSection />
     </>
   )
 }
