+++ src/sections/BioSection/BioSection.css
@@ -1,0 +1,7 @@
+.bio-section {
+    grid-column: span 4;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
