+++ src/sections/ProjectsSection/ProjectsSection.css
@@ -1,0 +1,7 @@
+.projects-section {
+    grid-column: span 4;
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    align-items: center;
+}
