+++ src/sections/ProjectsSection/ProjectsSection.jsx
@@ -1,0 +1,16 @@
+import { Grid } from "../../ui/Grid/Grid"
+import { HeaderTwo } from "../../ui/Typography/Typography";
+import "./ProjectsSection.css"
+
+export const ProjectsSection = () => {
+    return (
+        <Grid background="neutral">
+            <section className="projects-section">
+                <HeaderTwo>
+                    Featured Projects
+                </HeaderTwo>
+            </section>
+        </Grid>
+    )
+}
+
